[refactor]: alterações no projeto conforme pedido no PR.
