+++ src/LojaDoValdir/Models/Cliente.cs
@@ -1,4 +1,4 @@
-using LojaDoValdir.Models;
+namespace LojaDoValdir.Models;
 public class Cliente
 {
     public Guid Id { get; set; }
